moving pending_users to new class and file and updating tests
